requirements.txt
aniso8601==1.2.0
click==6.6
Flask==0.11.1
Flask-HTTPAuth==3.2.1
Flask-PyMongo==0.4.1
Flask-RESTful==0.3.5
Flask-Testing==0.6.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pymongo==3.4.0
python-dateutil==2.6.0
pytz==2016.7
six==1.10.0
Werkzeug==0.11.11
wheel==0.24.0
requests-2.12.3